Initial extensions to 3D
This fork is going to be aiming to extend the original RigidLibrary routines that allow for rigidity calculations in 2D to 3D.  This commit contains three important additions to the RigidLibrary that will be built on going forward:

Configuration_altered.py:  This contains an altered Configuration class, like the one found in Configuration.py.  The only difference is that this one is easier to load in contact networks into the Configuration class.

Hessian_3D.py:  This file contains beginning of functions that will ultimately construct 3D Hessians using Finn's calculation.

testing3DRigidity.py:  This contains practice lattices such as simple cubic, body centered cubic, and face centered cubic so that we can practice using our rigidity routines on them.