Gave up on defining FCC and BCC lattices
I was trying to code up the positions, radii, and contact maps for the FCC and BCC lattices but I can't quite get them right.  I gave up on getting them right because I realized that the pebble game and hessian analysis' should deem the SC lattice rigid, and therefore there's not an immediate need for the other more coordinated lattices.  I commented and moved the code defining FCC and BCC lattices to the end of testing3DRigidity.py with a comment saying that they need to be fixed and also what I think might be wrong with them.